Feat: cleaaning the bip boop slop into cleaner components
