+++ typescript/packages/x402/src/paywall/src/solana/types.ts
@@ -1,0 +1,6 @@
+import type { WalletWithSolanaFeatures } from "@solana/wallet-standard-features";
+
+export type WalletOption = {
+  value: string;
+  wallet: WalletWithSolanaFeatures;
+};
